% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bspline_basis}
\alias{bspline_basis}
\title{Create a new B-spline basis.}
\usage{
bspline_basis(xrange, nbasis, intercept, degree = 3)
}
\arguments{
\item{xrange}{The domain of the B-spline model.}

\item{nbasis}{The number of basis functions to use.}

\item{intercept}{Should the basis include an intercept?}

\item{degree}{The degree of the polynomials.}
}
\description{
Create a new B-spline basis.
}

