% Generated by roxygen2 (4.0.1): do not edit by hand
\name{curve_e_step}
\alias{curve_e_step}
\title{Compute sufficient statistics and likelihood.}
\usage{
curve_e_step(curve, model, ss)
}
\arguments{
\item{curve}{The curve to process.}

\item{model}{A trajclust model.}

\item{ss}{A trajclust sufficient statistics container.}
}
\description{
Compute sufficient statistics and likelihood.
}

