% Generated by roxygen2 (4.0.1): do not edit by hand
\name{curve_var_e_step}
\alias{curve_var_e_step}
\title{Compute sufficient statistics and likelihood using a variational
approximation.}
\usage{
curve_var_e_step(curve, model, ss)
}
\arguments{
\item{curve}{The curve to process.}

\item{model}{A trajclust model.}

\item{ss}{A trajclust sufficient statistics container.}
}
\description{
Compute sufficient statistics and likelihood using a variational
approximation.
}

