% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diagonal_covariance}
\alias{diagonal_covariance}
\title{Create a new diagonal covariance function.}
\usage{
diagonal_covariance(noise)
}
\arguments{
\item{noise}{The square root of the measurement variance.}
}
\description{
Create a new diagonal covariance function.
}

