% Generated by roxygen2 (4.0.1): do not edit by hand
\name{init_trajclust_model}
\alias{init_trajclust_model}
\title{Initialize a trajclust model using data.}
\usage{
init_trajclust_model(curveset, model)
}
\arguments{
\item{curveset}{A collection of curves.}

\item{model}{A trajclust model.}
}
\description{
Initialize a trajclust model using data.
}

