% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_curve}
\alias{make_curve}
\title{Create an individual curve from a data.frame.}
\usage{
make_curve(curve_data)
}
\arguments{
\item{curve_data}{A \code{data.frame} containing measurement times
\code{x}, measurements \code{y}, and a curve ID \code{id}.}
}
\description{
Create an individual curve from a data.frame.
}

