% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_curveset}
\alias{make_curveset}
\title{Create curves from "narrow" data.}
\usage{
make_curveset(x, y, curve_id)
}
\arguments{
\item{x}{A vector of measurement times.}

\item{y}{A vector of measurements corresponding to \code{x}.}

\item{curve_id}{Curve IDs for measurements in \code{x} and
\code{y}.}
}
\description{
Create curves from "narrow" data.
}

