% Generated by roxygen2 (4.0.1): do not edit by hand
\name{new_trajclust_model}
\alias{new_trajclust_model}
\title{Create an empty trajclust model.}
\usage{
new_trajclust_model(G, P, basis, covariance, bcov = NULL)
}
\arguments{
\item{G}{The number of clusters.}

\item{P}{The number of basis functions.}

\item{basis}{A function to compute design matrices.}

\item{covariance}{A function to compute covariance matrices.}

\item{bcov}{Individual offset initial covariance.}
}
\description{
Create an empty trajclust model.
}

