% Generated by roxygen2 (4.0.1): do not edit by hand
\name{new_trajclust_suffstats}
\alias{new_trajclust_suffstats}
\title{Create an empty set of trajclust sufficient statistics.}
\usage{
new_trajclust_suffstats(model)
}
\arguments{
\item{model}{A trajclust model.}
}
\description{
Create an empty set of trajclust sufficient statistics.
}

