% Generated by roxygen2 (4.0.1): do not edit by hand
\name{squared_exp_covariance}
\alias{squared_exp_covariance}
\title{Create a new squared exponential covariance function.}
\usage{
squared_exp_covariance(amp, bw, noise)
}
\arguments{
\item{amp}{The amplitude of the kernel.}

\item{bw}{The bandwidth of the kernel.}

\item{noise}{The amount of observation noise.}
}
\description{
Create a new squared exponential covariance function.
}

