% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{trajclust}
\alias{trajclust}
\alias{trajclust-package}
\title{trajclust.}
\usage{
trajclust(x, y, id, ngroups, xrange = range(x), nbasis, amp, bw, noise,
  bmean = NULL, bcov = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A vector of measurement times.}

\item{y}{A vector of measurement values.}

\item{id}{A vector of trajectory IDs for each (x, y) pair.}

\item{ngroups}{The number of clusters to estimate.}

\item{xrange}{The lower and upper bounds of the measurement times.}

\item{nbasis}{The number of basis functions to use.}

\item{amp}{The amplitude of the correlation model.}

\item{bw}{The bandwidth of the correlation model.}

\item{noise}{The observation standard deviation.}

\item{bmean}{Normal mean parameter for parametric random effects.}

\item{bcov}{Normal covariance parameter for parametric random effects.}

\item{verbose}{Logical flog indicating whether to print convergence
information.}
}
\description{
trajclust.

Cluster trajectories.
}

