% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trajclust_elbo}
\alias{trajclust_elbo}
\title{Compute the ELBO of the trajclust model with the given variational
distributions.}
\usage{
trajclust_elbo(X, x, y, K, var_z, var_bmean, var_bcov, model)
}
\arguments{
\item{X}{A design matrix for a curve.}

\item{x}{The measurement times for a curve.}

\item{y}{The observed measurements for a curve.}

\item{K}{The covariance matrix of the observed measurements.}

\item{var_z}{The variational distribution over groups.}

\item{var_bmean}{The variational mean over curve offsets.}

\item{var_bcov}{The variational covariance over curve offsets.}

\item{model}{A trajclust model.}
}
\description{
Compute the ELBO of the trajclust model with the given variational
distributions.
}

