% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trajclust_full_var_inference}
\alias{trajclust_full_var_inference}
\title{Use a trained model to infer groups and offsets and compute
likelihood.}
\usage{
trajclust_full_var_inference(curveset, model)
}
\arguments{
\item{curveset}{A collection of curves to evaluate.}

\item{model}{A trained trajclust model.}
}
\description{
Use a trained model to infer groups and offsets and compute
likelihood.
}

