% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trajclust_inference}
\alias{trajclust_inference}
\title{Compute posterior for a curve.}
\usage{
trajclust_inference(X, y, K, model)
}
\arguments{
\item{X}{A design matrix for the curve.}

\item{y}{A response vector for the curve.}

\item{K}{The covariance matrix of the responses.}

\item{model}{A trajclust model.}
}
\description{
Compute posterior for a curve.
}

