% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trajclust_mle}
\alias{trajclust_mle}
\title{Compute maximum likelihood estimate of trajclust model.}
\usage{
trajclust_mle(model, ss)
}
\arguments{
\item{model}{A trajclust model.}

\item{ss}{A trajclust sufficient statistics container.}
}
\description{
Compute maximum likelihood estimate of trajclust model.
}

