% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trajclust_var_inference}
\alias{trajclust_var_inference}
\title{Compute posterior group and offset for a curve.}
\usage{
trajclust_var_inference(X, x, y, K, model, tol = 1e-08)
}
\arguments{
\item{X}{The design matrix for the curve.}

\item{x}{The measurement times for the curve.}

\item{y}{The observed measurements for the curve.}

\item{K}{The measurement covariance matrix.}

\item{model}{A trajclust model.}

\item{tol}{Convergence tolerance for variational inference.}
}
\description{
Compute posterior group and offset for a curve.
}

