% Generated by roxygen2 (4.0.1): do not edit by hand
\name{new_trajclust_model}
\alias{new_trajclust_model}
\title{Create an empty trajclust model.}
\usage{
new_trajclust_model(k, p, basis, covariance)
}
\arguments{
\item{k}{The number of clusters.}

\item{p}{The number of basis functions.}

\item{basis}{A function to compute design matrices.}

\item{covariance}{A function to compute covariance matrices.}
}
\description{
Create an empty trajclust model.
}

