% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run_em}
\alias{run_em}
\title{Run EM to fit the trajclust model.}
\usage{
run_em(curveset, model, tol = 1e-08)
}
\arguments{
\item{curveset}{A collection of observed curves.}

\item{model}{A trajclust model.}

\item{tol}{Convergence tolerance.}
}
\description{
Run EM to fit the trajclust model.
}

