% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run_var_em}
\alias{run_var_em}
\title{Run variational EM to fit the trajclust model. The model must have
been initialized with \code{bmean} and \code{bcov}.}
\usage{
run_var_em(curveset, model, tol = 1e-08, maxiter = 10000, verbose = TRUE)
}
\arguments{
\item{curveset}{A collection of observed curves.}

\item{model}{A trajclust model.}

\item{tol}{Convergence tolerance.}

\item{maxiter}{The maximum number of EM iterations.}

\item{verbose}{Logical flag indicating whether to print convergence
information.}
}
\description{
This procedure is used to estimate the model when individual curves
can have linear offsets.
}

