% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{trajclust}
\alias{trajclust}
\alias{trajclust-package}
\title{trajclust.}
\usage{
trajclust(x, y, id, ngroups, xrange = range(x), nbasis, amb, bw,
  bcov = NULL, verbose = TRUE, maxiter = 1000, model = NULL)
}
\arguments{
\item{x}{A vector of measurement times.}

\item{y}{A vector of measurement values.}

\item{id}{A vector of trajectory IDs for each (x, y) pair.}

\item{ngroups}{The number of clusters to estimate.}

\item{xrange}{The lower and upper bounds of the measurement times.}

\item{nbasis}{The number of basis functions to use.}

\item{amp}{The amplitude of the correlation model.}

\item{bw}{The bandwidth of the correlation model.}

\item{bcov}{Initial normal covariance parameter for parametric random effects.}

\item{verbose}{Logical flog indicating whether to print convergence
information.}

\item{model}{Continue to fit this model. Providing this argument
causes most of the others to be ignored.}
}
\description{
trajclust.

Cluster trajectories.
}

